requirements.txt
langchain
langchain-chroma
langchain-huggingface
langchain-community
chromadb
huggingface_hub
transformers
torch
tabulate
pandas
numpy
python-dotenv
fastapi
uvicorn
pydantic
requests
sentence-transformers
pytest